+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Add new collision group implementation [#1091](https://github.com/excaliburjs/Excalibur/issues/1091) [#862](https://github.com/excaliburjs/Excalibur/issues/862)
 - New `ex.Collider` type which is the container for all collision related behavior and state. Actor is now extracted from collision.
 - Added interface `Clonable<T>` to indicate if an object contains a clone method
 - Added interface `Eventable<T>` to indicated if an object can emit and receive events
@@ -61,6 +62,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Deprecated
 
+- Legacy groups `ex.Group` will be removed in v0.24.0, use collision groups as a replacement [#1091](https://github.com/excaliburjs/Excalibur/issues/1091)
+- Legacy collision groups off `Actor` will be removed in v0.24.0, use `Actor.body.collider.collisionGroup` [#1091](https://github.com/excaliburjs/Excalibur/issues/1091)
 - Removed `NaiveCollisionBroadphase` as it was no longer used
 - Renamed methods and properties will be available until `v0.24.0`
 - Deprecated collision attributes on actor, use `Actor.body.collider`